Add initial XSLT and CSS stylesheets.
This makes it possible to view a pretty version of the task lists by
loading the XML file directly in the browser.  No scripts or external
programs are necessary; simply pointing a browser at the files in a git
repository will do.

Ideally, it would be possible to view these via cgit's tree/plain
feature as well.  Unfortunately, cgit.freedesktop.org doesn't appear to
return the correct mime-types (application/xml, application/xslt+xml,
and text/css) to make that happen.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>