gl33: Mark VertexAttribDivisorDefault as done.
Technically, we don't have a Piglit test for this.  But if it got set to
non-zero, virtually every normal drawing call would break.  The only
ways it could work are: a) the divisor is ignored completely (which
would cause the ARB_instanced_arrays tests to fail), or b) it's set to
some huge value larger than the number of primitives in any draw call.

Those seem really unlikely, so I think I'm satisfied with our test
coverage.